moving AddMemberModal to components
